refactor: restructure codebase with modular command and utility organization
