+++ src/commands/mod.rs
@@ -12,16 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod cli;
-pub mod commands;
-pub mod config;
-pub mod executor;
-pub mod node;
-pub mod ssh;
-pub mod ui;
-pub mod utils;
-
-pub use cli::Cli;
-pub use config::Config;
-pub use executor::ParallelExecutor;
-pub use node::Node;
+pub mod download;
+pub mod exec;
+pub mod list;
+pub mod ping;
+pub mod upload;
