+++ src/utils/mod.rs
@@ -12,16 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod cli;
-pub mod commands;
-pub mod config;
-pub mod executor;
-pub mod node;
-pub mod ssh;
-pub mod ui;
-pub mod utils;
+pub mod fs;
+pub mod logging;
+pub mod output;
 
-pub use cli::Cli;
-pub use config::Config;
-pub use executor::ParallelExecutor;
-pub use node::Node;
+pub use fs::{format_bytes, resolve_source_files, walk_directory};
+pub use logging::init_logging;
+pub use output::save_outputs_to_files;
